Added recursive `Let` primitive to simulate `Mu`.
In this commit:

- Let parsing
- Let evaluation
- Let test cases for both `eval` and `parse`

Examples of the let syntax are

```
let a t b
let a = t; b
let a(x : Foo, y : Bar) : Baz = t; b
...
```

Let mixes cleanly with Lam, and different shadowed names can be
referenced via carets `^x` which dereferences scope from inside to out.

Let currently has a `Recr` marker to indicate whether the recursive term
is normalizable or not, which may be useful for equirecursive types

Various examples of this have been added to `Spec.Lang` and `Spec.Core.`

In the `Equality` module are some simple examples converting the
recursive terms with `Mu` from `FirstOrder`, and a draft of a
first-order encoding function.