New equality algorithm
Works just as well as the current Formality JavaScript implementation of
equality (halting/looping in the exact same cases); but runs much more
efficiently.

The algorithm, however, is yet not aware of the regularity of second
order trees, as defined by Neal Glew, so it still loops in cases where
we could in principle conclude the equality of terms.